Feat: JWT at the Edge!
Prevent unauthorized access to the the API Gateway by issuing and validating JWT tokens.
    
If an API request comes from the domain (cloudfront distributionName), generate a short-lived JWT and set it as a cookie. If the cookie is present, validate the JWT cookie and allow access to api, otherwise return a 401 unauthorized.
    
    ToDo:
    - encrypt payload
    - only set cookie if not present
    - refresh cookie on expiration
    - delete cookie if invalid